+++ examples/settings.py
@@ -1,21 +1,17 @@
 #!/usr/bin/env python3
+import time
 
-from pylecroy.pylecroy import Lecroy
-from pylecroy.pylecroy import Channels
-
+from pylecroy.pylecroy import *
+import numpy as np
+import matplotlib.pyplot as plt
 import sys
 
 VERSION = '0.1.0'
 
-USAGE = '''
-Return information from Lecroy
-
+USAGE = '''basic: parameters settings usage
 Usage:
-    python lecroy_info.py -a "IP:10.67.16.22"
-    python lcry_info.py -a "IP:10.67.16.22"
-    python lcry_info.py -a "USBTMC:<Host Name>"
-    ...
-        
+    python basic.py -a "IP:10.67.16.22"
+
 Options:
     -h, --help              this help message.
     -v, --version           version info.
@@ -29,7 +25,7 @@ def main(argv=None):
     if argv is None:
         argv = sys.argv[1:]
     try:
-        opts, args = getopt.gnu_getopt(argv, 'hvla:', ['help', 'version', 'address='])
+        opts, args = getopt.gnu_getopt(argv, 'hva:', ['help', 'version', 'address='])
         address = None
 
         for o, a in opts:
@@ -57,11 +53,24 @@ def main(argv=None):
     scope = Lecroy(address)
 
     # Get scope parameters
-    print(f"scope identifier             : {scope.identifier}")
-    print(f"get trigger mode             : {scope.trigger_mode}")
-    print(f"get auto calibration         : {scope.auto_calibration}")
-    print(f"Display state                : {scope.display}")
-    print(f"Grid mode                    : {scope.grid}")
+    print(f"Identifier               : {scope.identifier}")
+
+    # Trigger setting
+    trigger = scope.trigger_mode
+    print(f"Trigger mode             : {trigger}")
+    for mode in TriggerModes.MODES:
+        scope.trigger_mode = mode
+        print(f"Trigger mode             : {scope.trigger_mode}")
+
+    # Auto calibration
+    cal = scope.auto_calibration
+    print(f"Auto calibration         : {cal}")
+    scope.auto_calibration = Calibration.OFF
+    print(f"Auto calibration         : {scope.auto_calibration}")
+    scope.auto_calibration = cal
+    print(f"Auto calibration         : {scope.auto_calibration}")
+
+    # Grid
 
     scope.close()